+++ lib/examples/iter2.rs
@@ -120,8 +120,7 @@ fn main() {
             },
             Ok(Item::Option(_, _))                  |   // All real ones covered above
             Ok(Item::EarlyTerminator)               |   // Do not need to know
-            Err(ProblemItem::LongMissingData(_))    |   // No data taking options!
-            Err(ProblemItem::ShortMissingData(_))   |   // No data taking options!
+            Err(ProblemItem::MissingOptionData(_))  |   // No data taking options!
             Ok(Item::Command(_))                    |   // No commands here!
             Err(ProblemItem::AmbiguousCmd(_))       |   // No commands here!
             Err(ProblemItem::UnknownCommand(_, _)) => { // No commands here!