+++ lib/examples/mining2.rs
@@ -110,8 +110,7 @@ fn main() {
                 eprintln!("Error: Option `{}` does not take data, but some was provided", opt);
             },
             // There are no more problems applicable to this example!
-            ProblemItem::LongMissingData(_)    |   // No data taking options!
-            ProblemItem::ShortMissingData(_)   |   // No data taking options!
+            ProblemItem::MissingOptionData(_)  |   // No data taking options!
             ProblemItem::AmbiguousCmd(_)       |   // No commands here!
             ProblemItem::UnknownCommand(_, _) => { // No commands here!
                 unreachable!();